requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.49.0
graphene==3.0
graphene-django==3.0.0b7
graphql-core==3.1.7
graphql-relay==3.1.5
idna==3.3
iso8601==1.0.2
oauthlib==3.2.0
promise==2.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
